Re-introduce Spring XML based configuration
This is to retain existing Spring XML based configuration in the
microservice and continue to support switching message bus
implementations through the existing beans XML file resource.

The Java based configuration, though feasible, introduced complexity.
We are going to take the pragmatic approach of handling topic password
de-obfuscation centrally in the event-client.

Change-Id: I1c0ed92cdc155a5d848f94dafedb6c69af88b681
Issue-ID: AAI-1388
Signed-off-by: Michael Arrastia <MArrasti@amdocs.com>